@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "nonebot_plugin_nowtime"
-version = "0.1.5"
+version = "0.2.0"
 description = "Nonebot2 plugin for telling the time per hour"
 authors = ["Cvandia"]
 license = "MIT"
@@ -9,18 +9,15 @@ homepage = "https://github.com/Cvandia/nonebot_plugin_nowtime"
 repository = "https://github.com/Cvandia/nonebot_plugin_nowtime"
 
 [tool.poetry.dependencies]
-python = ">=3.7"
+python = ">=3.9,<4.0"
 nonebot2 = ">=2.0.0-beta.1"
-nonebot-adapter-onebot = ">=2.0.0-beta.1"
+nonebot-adapter-onebot = ">=2.0.0rc1"
 httpx = ">=0.19.0"
 aiofiles = ">=0.7.0"
 nonebot-plugin-apscheduler= ">=0.1.3"
 pydantic = ">=1.10.2"
 pathlib = ">=1.0.1"
-
-[tool.poetry.dev-dependencies]
-nb-cli = "^0.6.0"
-black = "^22.1.0"
+ujson = ">=5.10.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
