@@ -1,7 +1,8 @@
 from nonebot.plugin import on_regex, PluginMetadata
 from nonebot.adapters.onebot.v11.permission import GROUP_ADMIN, GROUP_OWNER
 from nonebot.permission import SUPERUSER
-from nonebot.params import Matcher, RegexGroup
+from nonebot.params import RegexGroup
+from nonebot.matcher import Matcher
 from nonebot import require, get_bot, get_driver
 from nonebot.adapters.onebot.v11 import (
     Bot,
@@ -18,7 +19,6 @@ from datetime import datetime
 from .config import Config
 import asyncio
 import aiofiles
-import os
 try:
     import ujson as json
 except ModuleNotFoundError:
@@ -68,7 +68,7 @@ except Exception:
 
 time_now = on_regex(r"(现在|当前|北京)时间", block=True, priority=5)
 trun_on_nowtime = on_regex(r"^(开启|关闭)整点报时([0-9]*)$", priority=99,
-                           block=True, permission=SUPERUSER | GROUP_ADMIN | GROUP_OWNER)
+                        block=True, permission=SUPERUSER | GROUP_ADMIN | GROUP_OWNER)
 list_matcher = on_regex(r"^查看整点报时列表$", priority=99,
                         permission=SUPERUSER | GROUP_ADMIN | GROUP_OWNER)
 
@@ -80,9 +80,9 @@ async def _():
         get_json = await client.get('https://v.api.aa1.cn/api/time-tx/index.php', timeout=30)
     get_msg = json.loads(get_json.text)
     msg = (f"⭐{get_msg['msg']}⭐\n"
-           + f"\n现在是北京时间:\n{get_msg['nowtime']}"
-           + f"送你一句：\n⭐{get_msg['nxyj']}⭐"
-           )
+        + f"\n现在是北京时间:\n{get_msg['nowtime']}"
+        + f"送你一句：\n⭐{get_msg['nxyj']}⭐"
+        )
     await time_now.send(message=(msg))
 
 
@@ -131,14 +131,17 @@ async def _(
     await matcher.finish(f"已成功{mode}{group_id}的整点报时")
 
 # 读取配置的开始和结束时间
-star_time = Config.parse_obj(get_driver().config.dict()).start_time
-end_time = Config.parse_obj(get_driver().config.dict()).end_time
-
+try:
+    star_time = get_driver().config.start_time
+    end_time = get_driver().config.end_time
+except Exception:
+    star_time = 0
+    end_time = 24
 # 发送报时
 
 
 async def post_scheduler():
-    bot: Bot = get_bot()
+    bot: Bot = get_bot() #type: ignore
     delay = 2 * 0.5
     if datetime.now().hour in range(star_time, end_time):
         for group_id in CONFIG["opened_groups"]:
@@ -146,14 +149,14 @@ async def post_scheduler():
                 # 整点语音
                 url = 'https://v.api.aa1.cn/api/api-baoshi/data/baoshi/'
                 url = url + str(datetime.now().hour) + '.mp3'
-                record = MessageSegment.record(url)
-                await bot.send_group_msg(group_id=int(group_id), message=record)
+                record = MessageSegment.record(url, timeout=30)
+                await bot.send_group_msg(group_id=int(group_id), message=record) #type: ignore
             except ActionFailed as e:
                 logger.warning(f'{repr(e)}')
             await asyncio.sleep(delay)
             try:
                 msg = await get_word_result()
-                await bot.send_group_msg(group_id=int(group_id), message=msg)
+                await bot.send_group_msg(group_id=int(group_id), message=msg) #type: ignore
             except ActionFailed as e:
                 logger.warning(
                     f"定时发送整点报时到 {group_id} 失败，可能是风控或机器人不在该群聊 {repr(e)}")
@@ -161,13 +164,14 @@ async def post_scheduler():
 
 
 # 加载整点报时词库
-words_for_time = json.load(open(Path(os.path.join(os.path.dirname(
-    __file__), "resource")) / "time_words.json", "r", encoding="utf8"))
+words_for_time:Dict[str,str] = json.load(
+    open(Path(__file__).parent / 'resource/time_words.json', 'r', encoding='utf-8'))
 
 # 匹配词库
 
 
-async def get_word_result() -> str:
+async def get_word_result():
+    """匹配词库"""
     keys = words_for_time.keys()
     for key in keys:
         try:
@@ -179,8 +183,9 @@ async def get_word_result() -> str:
 
 # 添加定时任务
 try:
-    scheduler.add_job(
-        post_scheduler, "cron", hour='*', id="everyday_nowtime"
-    )
+    if scheduler:
+        scheduler.add_job(
+            post_scheduler, "cron", hour='*', id="everyday_nowtime"
+        )
 except ActionFailed as e:
     logger.warning(f"定时任务添加失败，{repr(e)}")
